app/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
SQLAlchemy==2.0.12
Authlib==1.2.0
python-dotenv==1.0.0
requests==2.30.0
Werkzeug==2.3.4
psycopg2-binary==2.9.6